refactor: Rename 'compute' layer to 'parse' layer
