+++ parse/state_extract_operation_type_test.go
@@ -1,4 +1,4 @@
-package compute
+package parse
 
 import (
 	"testing"
