+++ parse/state_preparing.go
@@ -1,4 +1,4 @@
-package compute
+package parse
 
 import (
 	"errors"
